Auto-merged by TNRP. Authorized for merge by min.wang@salesforce.com.

+++ sam-internal/k8s-in/slb/scripts/slb-cleanup-unknownpods.sh
@@ -1,7 +1,12 @@
 #!/bin/bash -xe
 while true
 do
-    # Cleanup unknown pods
-    kubectl get pods -n sam-system | grep slb | grep Unknown | sed 's/ .*//'g | xargs kubectl -n sam-system delete pod --grace-period=0 --force
+    unknownpodcount=`kubectl get pods -n sam-system | grep slb | grep Unknown |wc -l`
+
+    if [ $unknownpodcount -gt 0 ]
+    then
+        kubectl get pods -n sam-system | grep slb | grep Unknown | sed 's/ .*//'g | xargs kubectl -n sam-system delete pod --grace-period=0 --force
+    fi
+
     sleep 600
 done
